@@ -5,7 +5,8 @@ Spree::CheckoutController.class_eval do
   end
 
   rescue_from TaxCloud::Errors::ApiError do |exception|
-    flash[:error] = Spree.t(:address_verification_failed)
+    exception_message = exception.problem
+    flash[:error] = Spree.t(:address_verification_failed) + (exception_message ? ": #{exception_message}" : '')
     redirect_to checkout_state_path(:address)
   end
 end
