@@ -15,7 +15,7 @@ module SolidusTaxCloud
       end
 
       def run_migrations
-        run_migration = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask 'Would you like to run the migrations now? [Y/n]')
+        run_migration = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask('Would you like to run the migrations now? [Y/n]'))
         if run_migration
           run 'bundle exec rake db:migrate'
         else
