@@ -4,7 +4,7 @@ module SolidusTaxCloud
     isolate_namespace Spree
     engine_name 'solidus_tax_cloud'
 
-    config.autoload_paths += %W(#{config.root}/lib)
+    config.autoload_paths += %W[#{config.root}/lib]
 
     # use rspec for tests
     config.generators do |generator|
@@ -25,18 +25,24 @@ module SolidusTaxCloud
           'lib/assets/javascripts/spree/frontend/spree_tax_cloud.js',
           'lib/assets/stylesheets/spree/frontend/spree_tax_cloud.css'
         ]
-        Dir.glob(File.join(File.dirname(__FILE__), "../../app/**/frontend/*_decorator*.rb")) do |c|
+        Dir.glob(File.join(File.dirname(__FILE__), '../../app/**/frontend/*_decorator*.rb')) do |c|
           Rails.configuration.cache_classes ? require(c) : load(c)
         end
       end
     end
 
+    def self.backend_available?
+      @@frontend_available ||= ::Rails::Engine.subclasses.map(&:instance).map { |e| e.class.to_s }.include?('Spree::Backend::Engine')
+    end
+
+    paths['app/controllers'] << 'lib/controllers/admin' if backend_available?
+
     def self.frontend_available?
       @@frontend_available ||= ::Rails::Engine.subclasses.map(&:instance).map { |e| e.class.to_s }.include?('Spree::Frontend::Engine')
     end
 
-    if self.frontend_available?
-      paths["app/controllers"] << "lib/controllers/frontend"
+    if frontend_available?
+      paths['app/controllers'] << 'lib/controllers/frontend'
     end
 
     config.to_prepare &method(:activate).to_proc
